ppt_tool/requirements.txt
python-pptx>=0.6.23
google-generativeai>=0.3.0
pywin32>=306; sys_platform == 'win32'
python-dotenv>=1.0.0
